reimplements generatePalette using imagemagic downsampling
This method yields better results than the quantize library, but still has weaknesses. Since this one works by downsampling the image to a handful pf pixels, it's actually returning the average or possibly most prominent color of each section of the image. This means it ignores small pops of color which may be desirable for someone generating a palette. I think what might make more sense is to downsample to the 16 color giferizer image and then capture those 16 colors. From there we can analyze the 16 colors and pick the palette based on delta-e.